+++ src/components/Login.jsx
@@ -1,56 +1,103 @@
 import styled from '@emotion/styled';
 import React from 'react';
+import { Link } from 'react-router-dom';
 import { images } from '../assets/assets';
 import { COLORS, SIZES } from '../styles/constants';
 
-const Container = styled.div`
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
+function LoginPageHeader() {
+  const Logo = () => {
+    const Container = styled.div`
+      display: flex;
+      &:hover {
+        cursor: pointer;
+      }
+    `;
 
-  padding: 150px;
-  min-width: ${SIZES.HEADER_MIN_WIDTH};
-  position: sticky;
-  top: 0;
-  background-color: ${COLORS.WHITE};
-  font-weight: 700;
-`;
+    const Image = styled.img`
+      height: 60px;
 
-const InnerContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
+      @media only screen and (max-width: ${SIZES.HEADER_MIDDLE_WIDTH}) {
+        height: 40px;
+      }
+    `;
 
-  position: sticky;
-  top: 0;
-  background-color: ${COLORS.WHITE};
-  font-weight: 700;
+    return (
+      <Link to="/main">
+        <Container>
+          <Image src={images.logo} />
+        </Container>
+      </Link>
+    );
+  };
 
-  h1 {
-    font-size: 20px;
-    font-weight: 600;
-    color: ${COLORS.DARK_GRAY};
-    padding-bottom: 20px;
-  }
-`;
+  const Container = styled.header`
+    display: flex;
+    justify-content: center;
+    align-items: center;
 
-const Image = styled.img`
-  height: 50px;
-  padding-top: 20px;
+    padding: 10px;
+    border-bottom: 2px solid ${COLORS.PRIMARY};
+    min-width: ${SIZES.HEADER_MIN_WIDTH};
+    position: sticky;
+    top: 0;
+    background-color: white;
+    z-index: 1;
+  `;
 
-  &:hover {
-    opacity: 0.8;
-    cursor: pointer;
-  }
-  @media only screen and (max-width: ${SIZES.HEADER_MIDDLE_WIDTH}) {
-    height: 40px;
-  }
-`;
+  return (
+    <Container>
+      <Logo />
+    </Container>
+  );
+}
 
 function Login() {
+  const Container = styled.div`
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+
+    padding: 150px;
+    min-width: ${SIZES.HEADER_MIN_WIDTH};
+    position: sticky;
+    top: 0;
+    background-color: ${COLORS.WHITE};
+    font-weight: 700;
+  `;
+
+  const InnerContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+
+    position: sticky;
+    top: 0;
+    background-color: ${COLORS.WHITE};
+    font-weight: 700;
+
+    h1 {
+      font-size: 20px;
+      font-weight: 600;
+      color: ${COLORS.DARK_GRAY};
+      padding-bottom: 20px;
+    }
+  `;
+
+  const Image = styled.img`
+    height: 50px;
+    padding-top: 20px;
+
+    &:hover {
+      opacity: 0.8;
+      cursor: pointer;
+    }
+    @media only screen and (max-width: ${SIZES.HEADER_MIDDLE_WIDTH}) {
+      height: 40px;
+    }
+  `;
   return (
     <>
       <Container>
@@ -65,4 +112,4 @@ function Login() {
   );
 }
 
-export default Login;
+export { LoginPageHeader, Login };
