+++ src/pages/LoginPage.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import Main from '../components/Main';
-import LoginPageHeader from '../components/LoginPageHeader';
-import Login from '../components/Login';
+import { LoginPageHeader, Login } from '../components/Login';
 
 function LoginPage() {
   return (
